restful/requirements.txt
asgiref==3.2.3
coverage==5.0.3
Django==3.0.3
django-crispy-forms==1.8.1
django-filter==2.2.0
django-nose==1.4.6
djangorestframework==3.11.0
nose==1.3.7
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
